scheduling existing medications at launch of server and reworked scheduling code to allow for scheduling of a single medication at a time as well as a list
