+++ src/controllers/controller.js
@@ -61,17 +61,8 @@ export const addNewMedication = async (req, res) => {
         })
         .execPopulate();
 
-    //we need to fully populate user for schedule code to work
-    //this is expensive, but the only other option is rewrite schedule code
-    await user
-        .populate({
-            path: "medications",
-            populate: { path: "dosages", model: "Dosage" },
-        })
-        .execPopulate();
-
     try {
-        await scheduleNewMedication(user, newMedication);
+        await scheduleNewMedication(newMedication);
     } catch (err) {
         console.error(err);
         return res.status(500).json({
@@ -472,163 +463,5 @@ const getWeeklyOccurrences = (user) => {
     scheduledDays.forEach((day) => {
         day.sort((a, b) => b.occurrence.date - a.occurrence.date);
     });
-    return scheduledDays;
-};
-
-/*
- * Function returns an array of days indexed from 0 (startDay) to (endDay - 1)
- * Inside each array, there are arrays of all the medications that are scheduled to be
- * taken that day. Inside the each medication object within the array, there is an array
- * that has the specific DateTime the medication needs to be taken. It also has the
- * dosageId.
- */
-export const getScheduledDays = (user, startDate, endDate) => {
-    //set end date to next saturday
-    if (endDate == null) {
-        endDate = new Date();
-        let day = endDate.getDay();
-        //add remaining days in the week in millis
-        endDate = new Date(endDate.getTime() + (6 - day) * 24 * 3600 * 1000);
-    }
-    //set start date to last sunday (or today if today is sunday)
-    if (startDate == null) {
-        startDate = new Date();
-        let day = startDate.getDay();
-        //sunday is index 0 so subtract current day in millis
-        startDate = new Date(startDate.getTime() - day * 24 * 3600 * 1000);
-    }
-    let scheduledDays = [];
-
-    //ceil both startDate and endDate to the endOfDay (avoids indexing issues later)
-    startDate = new Date(startDate.getTime());
-    startDate.setHours(23, 59, 59, 999);
-    endDate = new Date(endDate.getTime());
-    endDate.setHours(23, 59, 59, 999);
-
-    //get number of milliseconds between start date and end date
-    let days = endDate.getTime() - startDate.getTime();
-    //divide by number of milliseconds in one day and ceil
-    days = Math.ceil(days / (1000 * 3600 * 24));
-    //create entries in scheduledDays array for each day inbetween
-    for (let x = 0; x <= days; x++) {
-        scheduledDays.push([]);
-    }
-    if (user.medications == null) return scheduledDays;
-    user.medications.forEach((med) => {
-        let start = med.dateAdded;
-        //get number of milliseconds between start date and end date
-        let daysbetween = endDate.getTime() - start.getTime();
-        //divide by number of milliseconds in one day and ceil
-        daysbetween = Math.ceil(daysbetween / (1000 * 3600 * 24));
-        //start may not be startDate so create an offset
-        let offset = 0;
-        offset = startDate.getTime() - start.getTime();
-        offset = Math.floor(offset / (1000 * 3600 * 24));
-        //loop over days from start and only include them if they are after startDate
-        if (med.frequency.intervalUnit == "days") {
-            for (let i = 0; i < daysbetween; i += med.frequency.interval) {
-                if (daysbetween - i > days + 1) {
-                    //day is before start so just continue looping
-                    continue;
-                }
-                let daysbetween_m = i * (1000 * 3600 * 24);
-                //find actual date to take medication
-                let dateToTake = new Date(start.getTime() + daysbetween_m);
-                //round down to 12 am
-                dateToTake.setHours(0, 0, 0, 0);
-                let datesWTime = [];
-                med.dosages.forEach((dosage) => {
-                    //get amount of millis to add to current dateToTake
-                    let millisToAdd =
-                        dosage.reminderTime.getHours() * 3600 * 1000;
-                    let dateWTime = new Date(
-                        dateToTake.getTime() + millisToAdd
-                    );
-                    datesWTime.push({ date: dateWTime, dosageId: dosage._id });
-                });
-                if (datesWTime.length > 0) {
-                    scheduledDays[i - offset].push({
-                        datesWTime,
-                        medicationId: med._id,
-                    });
-                }
-            }
-        } else {
-            //if the start of each week isn't Sunday, make it Sunday
-            let days_s = start.getDay();
-            if (days_s != 0) {
-                start = new Date(start.getTime() - days_s * (1000 * 3600 * 24));
-            }
-
-            //loop over weeks so multiply interval by 7
-            for (let i = 0; i < daysbetween; i += 7 * med.frequency.interval) {
-                if (daysbetween - i - 6 > days + 1) {
-                    //day is before start so continue looping
-                    continue;
-                }
-                //loop over each day of this week
-                for (let j = i; j < i + 7; j++) {
-                    let daysbetween_m = (j - i) * (1000 * 3600 * 24);
-                    let dateToTake = new Date(start.getTime() + daysbetween_m);
-                    dateToTake.setHours(23, 59, 59, 999);
-
-                    //this kind of sucks but we need to check if the weekday matches the current day
-                    //and the current weekday is set to true in the database for this med
-                    //we also need to make sure the date is in the day range too
-                    if (
-                        (med.frequency.weekdays.sunday &&
-                            dateToTake.getDay() == 0) ||
-                        (med.frequency.weekdays.monday &&
-                            dateToTake.getDay() == 1) ||
-                        (med.frequency.weekdays.tuesday &&
-                            dateToTake.getDay() == 2) ||
-                        (med.frequency.weekdays.wednesday &&
-                            dateToTake.getDay() == 3) ||
-                        (med.frequency.weekdays.thursday &&
-                            dateToTake.getDay() == 4) ||
-                        (med.frequency.weekdays.friday &&
-                            dateToTake.getDay() == 5) ||
-                        (med.frequency.weekdays.saturday &&
-                            dateToTake.getDay() == 6 &&
-                            daysbetween - j <= days)
-                    ) {
-                        dateToTake.setHours(0, 0, 0, 0);
-                        let datesWTime = [];
-                        med.dosages.forEach((dosage) => {
-                            //get amount of millis to add to current dateToTake
-                            let millisToAdd =
-                                dosage.reminderTime.getHours() * 3600 * 1000;
-                            let dateWTime = new Date(
-                                dateToTake.getTime() + millisToAdd
-                            );
-                            datesWTime.push({
-                                date: dateWTime,
-                                dosageId: dosage._id,
-                            });
-                        });
-                        if (datesWTime.length > 0) {
-                            scheduledDays[j - i].push({
-                                datesWTime,
-                                medicationId: med._id,
-                            });
-                        }
-                    }
-                }
-            }
-        }
-    });
-    //eaving prints for testing purposes
-    // let i = 0;
-    // scheduledDays.forEach(day => {
-    //     console.log("day: " + i);
-    //     day.forEach(date => {
-    //         console.log(date.medicationId);
-    //         date.datesWTime.forEach(dateWTime => {
-    //             console.log(dateWTime.date.toString());
-    //             console.log(dateWTime.dosageId);
-    //         });
-    //     });
-    //     i++;
-    // });
     return scheduledDays;
 };
